Add files via upload
This project focuses on analyzing the performance of Liverpool FC players during the 2023-24 season, with an emphasis on their contributions in terms of goals and assists. The analysis is based on the number of matches played and the starting lineup. The key aspects of the study include:

Defenders: Analyzing goals and assists contributed by defenders.
Midfielders: Reviewing the performance of midfielders in terms of goals and assists.
Forwards: Assessing the offensive impact of forwards based on their goals and assists.
Goal-less and Assist-less Players: Identifying forwards and midfielders who have yet to score or assist, emphasizing those with no goal contributions.
This data-driven analysis is conducted using SQL queries, grouping the results by player names and sorting by assists in descending order. It aims to provide insights into the offensive contributions of Liverpool players during the 2023-24 season.